This PR updates the version of Aztec to incorporate a significant performance improvement for Gutenberg.

Related PRs:
- https://github.com/wordpress-mobile/AztecEditor-Android/pull/983
- https://github.com/WordPress/gutenberg/pull/40894
- https://github.com/wordpress-mobile/gutenberg-mobile/pull/4837

## To Test

### Gutenberg
This is a bit tricky to test since the only change here is that performance is improved.

One way to test this is to compare the a device's HWUI flamegraph with and without this change applied. For example, there should be a significant reduction in the number of slow frame "spikes" when scrolling through a post with 200 paragraph blocks (copying and pasting within the code editor is your friend here 😉 ). Below is what I observed when I did this test both with and without this change using release builds on my Pixel 3.

#### Before

Total frames rendered: 211
Janky frames: 9 (4.27%)
Janky frames (legacy): 62 (29.38%)
50th percentile: 15ms
90th percentile: 20ms
95th percentile: 23ms
99th percentile: 48ms

https://user-images.githubusercontent.com/4656348/167182465-96cddb99-38a6-40f8-9b85-d80e8033b456.mp4

#### After

Total frames rendered: 176
Janky frames: 0 (0.00%)
Janky frames (legacy): 2 (1.14%)
50th percentile: 8ms
90th percentile: 13ms
95th percentile: 15ms
99th percentile: 17ms

https://user-images.githubusercontent.com/4656348/167182511-5e8687d2-74aa-4c33-847d-27b802438aa5.mp4


### Aztec
1. Activate the classic editor on a self-hosted site: `Site Settings -> Use Block Editor`. Note that this can be difficult to do on account of https://github.com/wordpress-mobile/WordPress-Android/issues/14782, but with some persistence I was able to get it to work.
2. Open a new post with the Classic Editor
3. Add an image to the post
4. Add a video to the post
5. Observe that both the video and the image appear in the editor as expected.

## Regression Notes
1. Potential unintended areas of impact
The addition of images or videos within the Aztec editor could (but shouldn't) be impacted.

6. What I did to test those areas of impact (or what existing automated tests I relied on)
Manual testing.

7. What automated tests I added (or what prevented me from doing so)
No automated tests added as we have deprecated the Aztec Editor.

PR submission checklist:

- [X] I have completed the Regression Notes.
- [X] I have considered adding accessibility improvements for my changes.
- [x] I have considered if this change warrants user-facing release notes and have added them to `RELEASE-NOTES.txt` if necessary.